move common fixtures utils to utils.py and conftest
