+++ tests/integration/test_writes/test_partitioned_writes.py
@@ -51,91 +51,10 @@ from pyiceberg.types import (
     TimestamptzType,
 )
 
-TEST_DATA_WITH_NULL = {
-    'bool': [False, None, True],
-    'string': ['a', None, 'z'],
-    # Go over the 16 bytes to kick in truncation
-    'string_long': ['a' * 22, None, 'z' * 22],
-    'int': [1, None, 9],
-    'long': [1, None, 9],
-    'float': [0.0, None, 0.9],
-    'double': [0.0, None, 0.9],
-    'timestamp': [datetime(2023, 1, 1, 19, 25, 00), None, datetime(2023, 3, 1, 19, 25, 00)],
-    'timestamptz': [
-        datetime(2023, 1, 1, 19, 25, 00, tzinfo=pytz.timezone('America/New_York')),
-        None,
-        datetime(2023, 3, 1, 19, 25, 00, tzinfo=pytz.timezone('America/New_York')),
-    ],
-    'date': [date(2023, 1, 1), None, date(2023, 3, 1)],
-    # Not supported by Spark
-    # 'time': [time(1, 22, 0), None, time(19, 25, 0)],
-    # Not natively supported by Arrow
-    # 'uuid': [uuid.UUID('00000000-0000-0000-0000-000000000000').bytes, None, uuid.UUID('11111111-1111-1111-1111-111111111111').bytes],
-    'binary': [b'\01', None, b'\22'],
-    'fixed': [
-        uuid.UUID('00000000-0000-0000-0000-000000000000').bytes,
-        None,
-        uuid.UUID('11111111-1111-1111-1111-111111111111').bytes,
-    ],
-}
+from utils import TEST_DATA_WITH_NULL, TABLE_SCHEMA
 
 
-TABLE_SCHEMA = Schema(
-    NestedField(field_id=1, name="bool", field_type=BooleanType(), required=False),
-    NestedField(field_id=2, name="string", field_type=StringType(), required=False),
-    NestedField(field_id=3, name="string_long", field_type=StringType(), required=False),
-    NestedField(field_id=4, name="int", field_type=IntegerType(), required=False),
-    NestedField(field_id=5, name="long", field_type=LongType(), required=False),
-    NestedField(field_id=6, name="float", field_type=FloatType(), required=False),
-    NestedField(field_id=7, name="double", field_type=DoubleType(), required=False),
-    NestedField(field_id=8, name="timestamp", field_type=TimestampType(), required=False),
-    NestedField(field_id=9, name="timestamptz", field_type=TimestamptzType(), required=False),
-    NestedField(field_id=10, name="date", field_type=DateType(), required=False),
-    # NestedField(field_id=11, name="time", field_type=TimeType(), required=False),
-    # NestedField(field_id=12, name="uuid", field_type=UuidType(), required=False),
-    NestedField(field_id=11, name="binary", field_type=BinaryType(), required=False),
-    NestedField(field_id=12, name="fixed", field_type=FixedType(16), required=False),
-)
 
-
-@pytest.fixture(scope="session")
-def pa_schema() -> pa.Schema:
-    return pa.schema([
-        ("bool", pa.bool_()),
-        ("string", pa.string()),
-        ("string_long", pa.string()),
-        ("int", pa.int32()),
-        ("long", pa.int64()),
-        ("float", pa.float32()),
-        ("double", pa.float64()),
-        ("timestamp", pa.timestamp(unit="us")),
-        ("timestamptz", pa.timestamp(unit="us", tz="UTC")),
-        ("date", pa.date32()),
-        # Not supported by Spark
-        # ("time", pa.time64("us")),
-        # Not natively supported by Arrow
-        # ("uuid", pa.fixed(16)),
-        ("binary", pa.large_binary()),
-        ("fixed", pa.binary(16)),
-    ])
-
-
-@pytest.fixture(scope="session")
-def arrow_table_with_null(pa_schema: Schema) -> pa.Table:
-    """PyArrow table with all kinds of columns"""
-    return pa.Table.from_pydict(TEST_DATA_WITH_NULL, schema=pa_schema)
-
-
-@pytest.fixture(scope="session")
-def arrow_table_without_data(pa_schema: pa.Schema) -> pa.Table:
-    """PyArrow table with all kinds of columns"""
-    return pa.Table.from_pylist([], schema=pa_schema)
-
-
-@pytest.fixture(scope="session")
-def arrow_table_with_only_nulls(pa_schema: pa.Schema) -> pa.Table:
-    """PyArrow table with all kinds of columns"""
-    return pa.Table.from_pylist([{}, {}], schema=pa_schema)
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -280,8 +199,8 @@ def table_v1_appended_with_null_partitioned(session_catalog: Catalog, arrow_tabl
             properties={'format-version': '1'},
         )
 
-        for _ in range(2):
-            tbl.append(arrow_table_with_null)
+        tbl.append(arrow_table_with_null)
+        tbl.append(pa.concat_tables([arrow_table_with_null, arrow_table_with_null]))
         assert tbl.format_version == 1, f"Expected v1, got: v{tbl.format_version}"
 
 
@@ -427,8 +346,9 @@ def table_v2_appended_with_null_partitioned(session_catalog: Catalog, arrow_tabl
             properties={'format-version': '2'},
         )
 
-        for _ in range(2):
-            tbl.append(arrow_table_with_null)
+        tbl.append(arrow_table_with_null)
+        tbl.append(pa.concat_tables([arrow_table_with_null, arrow_table_with_null]))
+
         assert tbl.format_version == 2, f"Expected v2, got: v{tbl.format_version}"
 
 
@@ -519,7 +439,7 @@ def test_query_filter_only_nulls_partitioned(spark: SparkSession, part_col: str,
 
 @pytest.mark.integration
 @pytest.mark.parametrize(
-    "part_col", ['int', 'bool', 'string', "string_long", "long", "float", "double", "date", "timestamptz", "timestamp", "binary"]
+    "part_col", ['int',]# 'bool', 'string', "string_long", "long", "float", "double", "date", "timestamptz", "timestamp", "binary"]
 )
 @pytest.mark.parametrize("format_version", [1, 2])
 def test_query_filter_appended_null_partitioned(spark: SparkSession, part_col: str, format_version: int) -> None:
@@ -527,8 +447,10 @@ def test_query_filter_appended_null_partitioned(spark: SparkSession, part_col: s
     df = spark.table(identifier)
     for col in TEST_DATA_WITH_NULL.keys():
         df = spark.table(identifier)
-        assert df.where(f"{col} is not null").count() == 4, f"Expected 4 non-null rows for {col}"
-        assert df.where(f"{col} is null").count() == 2, f"Expected 2 null rows for {col}"
+        assert df.where(f"{col} is not null").count() == 6, f"Expected 6 non-null rows for {col}"
+        assert df.where(f"{col} is null").count() == 3, f"Expected 3 null rows for {col}"
+    rows = spark.sql(f"select partition from {identifier}.files").collect()
+    assert len(rows) == 6
 
 
 @pytest.mark.integration
@@ -544,6 +466,7 @@ def test_query_filter_v1_v2_append_null(spark: SparkSession, part_col: str) -> N
         assert df.where(f"{col} is null").count() == 2, f"Expected 2 null rows for {col}"
 
 
+@pytest.mark.integration
 def test_summaries_with_null(spark: SparkSession, session_catalog: Catalog, arrow_table_with_null: pa.Table) -> None:
     identifier = "default.arrow_table_summaries"
